Move importers to related source module
The logic of importing the CH dataset from csv files used to be in the root
app folder and this moves it to its own db source folder.
